fix (refs T33506): Move foundry-factories to a folder named "Factory" to improve readability of data structure
+adjust folder where further Factories will be created on using related command.
